Refactoring, Commenting, and Unit Tests Revisions for the Week of May 10, 2023 (#10)
* Created REST API Shell with Experts Endpoint for Search and Collaborators

This commit adds a temp

* Updated Project for Swagger UI in All Environments and Hosting

* Added API Pagination Interfaces for Getting Experts and Collaborators

* Lucene Filter Query Parameter for Experts

* Added Covid Experts for M2 Presentation

* Basic Elasticsearch Querying to Localhost

* Elasticsearch Environment Settings

This commit refactors the code to get the Elasticsearch URL, username, and password from configuration files so that this can be retrieved in production as an environment variable.

* Removing Unnecessary Files

This commit removes unnecessary bin and obj files from the repository and the gitignore will not add them later.

* Basic Publication and Relevancy Sum Metrics

This commit adds basic relevancy and publication count sums.

* Configurable Elasticsearch Query Limit

This commit adds a configurable value for the number of records to return from Elasticsearch queries.

* Added Configurable Application Insights Logging

This commit adds Application Insights logging via configurations in the app settings.

* Added Pagination to Experts Querying

This commit adds pagination to the expert search using query parameters.

* Added CI/CD via GitHub Actions

* Corrected Page Sorting Orders

This commit corrects an issue where the page sorting was in ascending order by default instead of descending.

* Fixed Null Name Bug

This commit fixes a bug where null names would end up being pooled as the same aggregation metric.

* Added API Setup Instructions to the README File

* Added API URL to the README

* Added Ascending/Descending Sorting for Getting Experts by Keywords

This commit adds ascending and descending sorts to the GET experts by keywords endpoint.

* Added Collaborators and Time Range Querying

This commit adds functionality to get a user's collaborator set and allows the request to be filtered within a time range.

* Added Citations Count in Expert Metrics

This commit adds citations count to the expert metrics. This looks at each expert's publications and then determines how many times all of their publications are cited by other experts within the context of a keywords search. It also provide the ability to sort the results by citations count (and changes citations count to be the default sorting metric).

* Fixed Issue of Non-Unique Collaborators

This commit fixes an issue where the collaborators listed for given author may not be unique by overriding the equality and hash code operators of the Expert class.

* Refactored Elasticsearch Functionality

This commit refactors some of the Elasticsearch querying functionality and sets up all of the new dependencies for injection.

* Revised API Setup In Instructions in the README

This commit revises the current setup instructions for the API and provides links of where to go to get various installations.

* Refactored and Commented the Experts Search Code

This commit adds comments to most of the code used for experts searching along with some minor refactoring as needed for clarity.

* Removed Used Connection String from App Insights Configuration

This commit removed the unused connection string setting from the AppliationInsights sections of the appsettings.json.

* Update .suo

* Elasticsearch Publication Search Unit Tests

This commit adds some basic randomized unit tests for the Elasticsearch publication search implementations.